fix toolbar, delivery screen, change to tablayout

+++ app/src/main/java/com/misa/fresher/model/Product.kt
@@ -3,6 +3,6 @@ package com.misa.fresher.model
 data class Product(
     var name: String,
     var id: String,
-    var price: Float,
-    var maxPrice: Float? = null
+    var price: Double,
+    var maxPrice: Double? = null
 )
